Move findMapPaths to directory.go
