+++ ygen/directory_test.go
@@ -399,3 +399,360 @@ func TestFindSchemaPath(t *testing.T) {
 		})
 	}
 }
+
+// TestFindMapPaths ensures that the schema paths that an entity should be
+// mapped to are properly extracted from a schema element.
+func TestFindMapPaths(t *testing.T) {
+	tests := []struct {
+		name                string
+		inStruct            *Directory
+		inField             string
+		inCompressPaths     bool
+		inShadowSchemaPaths bool
+		inAbsolutePaths     bool
+		wantPaths           [][]string
+		wantErr             bool
+	}{{
+		name: "first-level container with path compression off",
+		inStruct: &Directory{
+			Name: "AContainer",
+			Path: []string{"", "a-module", "a-container"},
+			Fields: map[string]*yang.Entry{
+				"field-a": {
+					Name: "field-a",
+					Parent: &yang.Entry{
+						Name: "a-container",
+						Parent: &yang.Entry{
+							Name: "a-module",
+						},
+					},
+				},
+			},
+		},
+		inField:   "field-a",
+		wantPaths: [][]string{{"field-a"}},
+	}, {
+		name: "invalid parent path - shorter than directory path",
+		inStruct: &Directory{
+			Name: "AContainer",
+			Path: []string{"", "a-module", "a-container"},
+			Fields: map[string]*yang.Entry{
+				"field-q": {
+					Name: "field-q",
+					Parent: &yang.Entry{
+						Name: "q-container",
+					},
+				},
+			},
+		},
+		inField: "field-q",
+		wantErr: true,
+	}, {
+		name: "first-level container with path compression on",
+		inStruct: &Directory{
+			Name: "BContainer",
+			Path: []string{"", "a-module", "b-container"},
+			Fields: map[string]*yang.Entry{
+				"field-b": {
+					Name: "field-b",
+					Parent: &yang.Entry{
+						Name: "config",
+						Parent: &yang.Entry{
+							Name: "b-container",
+							Parent: &yang.Entry{
+								Name: "a-module",
+							},
+						},
+					},
+				},
+			},
+			ShadowedFields: map[string]*yang.Entry{
+				"field-b": {
+					Name: "field-b",
+					Parent: &yang.Entry{
+						Name: "state",
+						Parent: &yang.Entry{
+							Name: "b-container",
+							Parent: &yang.Entry{
+								Name: "a-module",
+							},
+						},
+					},
+				},
+			},
+		},
+		inField:         "field-b",
+		inCompressPaths: true,
+		wantPaths:       [][]string{{"config", "field-b"}},
+	}, {
+		name: "first-level container with path compression on and ignoreShadowSchemaPaths on",
+		inStruct: &Directory{
+			Name: "BContainer",
+			Path: []string{"", "a-module", "b-container"},
+			Fields: map[string]*yang.Entry{
+				"field-b": {
+					Name: "field-b",
+					Parent: &yang.Entry{
+						Name: "config",
+						Parent: &yang.Entry{
+							Name: "b-container",
+							Parent: &yang.Entry{
+								Name: "a-module",
+							},
+						},
+					},
+				},
+			},
+			ShadowedFields: map[string]*yang.Entry{
+				"field-b": {
+					Name: "field-b",
+					Parent: &yang.Entry{
+						Name: "state",
+						Parent: &yang.Entry{
+							Name: "b-container",
+							Parent: &yang.Entry{
+								Name: "a-module",
+							},
+						},
+					},
+				},
+			},
+		},
+		inField:             "field-b",
+		inCompressPaths:     true,
+		inShadowSchemaPaths: true,
+		wantPaths:           [][]string{{"state", "field-b"}},
+	}, {
+		name: "container with absolute paths on",
+		inStruct: &Directory{
+			Name: "BContainer",
+			Path: []string{"", "a-module", "b-container", "c-container"},
+			Fields: map[string]*yang.Entry{
+				"field-d": {
+					Name: "field-d",
+					Parent: &yang.Entry{
+						Name: "c-container",
+						Parent: &yang.Entry{
+							Name: "b-container",
+							Parent: &yang.Entry{
+								Name: "a-module",
+							},
+						},
+					},
+				},
+			},
+		},
+		inField:         "field-d",
+		inAbsolutePaths: true,
+		wantPaths:       [][]string{{"", "b-container", "c-container", "field-d"}},
+	}, {
+		name: "top-level module - not valid to map",
+		inStruct: &Directory{
+			Name:   "CContainer",
+			Path:   []string{"", "c-container"}, // Does not have a valid module.
+			Fields: map[string]*yang.Entry{"top": {}},
+		},
+		inField: "top",
+		wantErr: true,
+	}, {
+		name: "list with leafref key",
+		inStruct: &Directory{
+			Name: "DList",
+			Path: []string{"", "d-module", "d-container", "d-list"},
+			ListAttr: &YangListAttr{
+				KeyElems: []*yang.Entry{
+					{
+						Name: "d-key",
+						Type: &yang.YangType{
+							Kind: yang.Yleafref,
+						},
+						Parent: &yang.Entry{
+							Name: "config",
+							Parent: &yang.Entry{
+								Name: "d-list",
+								Kind: yang.DirectoryEntry,
+								Dir: map[string]*yang.Entry{
+									"d-key": {
+										Name: "d-key",
+										Type: &yang.YangType{Kind: yang.Yleafref},
+									},
+								},
+								Parent: &yang.Entry{
+									Name: "d-container",
+									Parent: &yang.Entry{
+										Name: "d-module",
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			Fields: map[string]*yang.Entry{
+				"d-key": {
+					Name: "d-key",
+					Type: &yang.YangType{
+						Kind: yang.Yleafref,
+					},
+					Parent: &yang.Entry{
+						Name: "config",
+						Parent: &yang.Entry{
+							Name: "d-list",
+							Kind: yang.DirectoryEntry,
+							Dir: map[string]*yang.Entry{
+								"d-key": {
+									Name: "d-key",
+									Type: &yang.YangType{Kind: yang.Yleafref},
+								},
+							},
+							Parent: &yang.Entry{
+								Name: "d-container",
+								Parent: &yang.Entry{
+									Name: "d-module",
+								},
+							},
+						},
+					},
+				},
+			},
+			ShadowedFields: map[string]*yang.Entry{
+				"d-key": {
+					Name: "d-key",
+					Parent: &yang.Entry{
+						Name: "state",
+						Parent: &yang.Entry{
+							Name: "d-list",
+							Kind: yang.DirectoryEntry,
+							Dir: map[string]*yang.Entry{
+								"d-key": {
+									Name: "d-key",
+									Type: &yang.YangType{Kind: yang.Yleafref},
+								},
+							},
+							Parent: &yang.Entry{
+								Name: "d-container",
+								Parent: &yang.Entry{
+									Name: "d-module",
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		inField:         "d-key",
+		inCompressPaths: true,
+		wantPaths: [][]string{
+			{"config", "d-key"},
+			{"d-key"},
+		},
+	}, {
+		name: "list with leafref key with shadowSchemaPaths=true",
+		inStruct: &Directory{
+			Name: "DList",
+			Path: []string{"", "d-module", "d-container", "d-list"},
+			ListAttr: &YangListAttr{
+				KeyElems: []*yang.Entry{
+					{
+						Name: "d-key",
+						Type: &yang.YangType{
+							Kind: yang.Yleafref,
+						},
+						Parent: &yang.Entry{
+							Name: "config",
+							Parent: &yang.Entry{
+								Name: "d-list",
+								Kind: yang.DirectoryEntry,
+								Dir: map[string]*yang.Entry{
+									"d-key": {
+										Name: "d-key",
+										Type: &yang.YangType{Kind: yang.Yleafref},
+									},
+								},
+								Parent: &yang.Entry{
+									Name: "d-container",
+									Parent: &yang.Entry{
+										Name: "d-module",
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+			Fields: map[string]*yang.Entry{
+				"d-key": {
+					Name: "d-key",
+					Type: &yang.YangType{
+						Kind: yang.Yleafref,
+					},
+					Parent: &yang.Entry{
+						Name: "config",
+						Parent: &yang.Entry{
+							Name: "d-list",
+							Kind: yang.DirectoryEntry,
+							Dir: map[string]*yang.Entry{
+								"d-key": {
+									Name: "d-key",
+									Type: &yang.YangType{Kind: yang.Yleafref},
+								},
+							},
+							Parent: &yang.Entry{
+								Name: "d-container",
+								Parent: &yang.Entry{
+									Name: "d-module",
+								},
+							},
+						},
+					},
+				},
+			},
+			ShadowedFields: map[string]*yang.Entry{
+				"d-key": {
+					Name: "d-key",
+					Parent: &yang.Entry{
+						Name: "state",
+						Parent: &yang.Entry{
+							Name: "d-list",
+							Kind: yang.DirectoryEntry,
+							Dir: map[string]*yang.Entry{
+								"d-key": {
+									Name: "d-key",
+									Type: &yang.YangType{Kind: yang.Yleafref},
+								},
+							},
+							Parent: &yang.Entry{
+								Name: "d-container",
+								Parent: &yang.Entry{
+									Name: "d-module",
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		inField:             "d-key",
+		inCompressPaths:     true,
+		inShadowSchemaPaths: true,
+		wantPaths: [][]string{
+			{"state", "d-key"},
+			{"d-key"},
+		},
+	}}
+
+	for _, tt := range tests {
+		got, err := findMapPaths(tt.inStruct, tt.inField, tt.inCompressPaths, tt.inShadowSchemaPaths, tt.inAbsolutePaths)
+		if err != nil {
+			if !tt.wantErr {
+				t.Errorf("%s: YANGCodeGenerator.findMapPaths(%v, %v): compress: %v, shadowSchemaPaths: %v, got unexpected error: %v",
+					tt.name, tt.inStruct, tt.inField, tt.inCompressPaths, tt.inShadowSchemaPaths, err)
+			}
+			continue
+		}
+
+		if diff := cmp.Diff(tt.wantPaths, got); diff != "" {
+			t.Errorf("%s: YANGCodeGenerator.findMapPaths(%v, %v): compress: %v, shadowSchemaPaths: %v, (-want, +got):\n%s", tt.name, tt.inStruct, tt.inField, tt.inCompressPaths, tt.inShadowSchemaPaths, diff)
+		}
+	}
+}
